requirements.txt
Flask==2.2.2
Flask-RESTful==0.3.9
python-dotenv==0.19.2
pymongo==3.12.1
Flask-Cors==3.1.1
